# public/openapi.yaml
openapi: 3.0.3
info:
  title: ApplyCopilot REST API (Phase 1)
  description: API Contracts for Authentication and CV Import pipelines in ApplyCopilot V2.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/admin/llm-config:
    get:
      summary: Retrieve LLM provider configurations
      description: Returns the current LLM provider configuration and credential status. ADMIN role required.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Configuration retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                    properties:
                      defaultProvider:
                        type: string
                        enum: [ollama, gemini, claude]
                      parsingProvider:
                        type: string
                        enum: [ollama, gemini, claude]
                      summariesProvider:
                        type: string
                        enum: [ollama, gemini, claude]
                  credentialStatus:
                    type: object
                    properties:
                      ollama:
                        type: boolean
                      gemini:
                        type: boolean
                      claude:
                        type: boolean
        '401':
          description: Unauthorized. Active session cookie required.
        '403':
          description: Forbidden. ADMIN role required.
    post:
      summary: Update LLM provider configurations
      description: Updates the LLM provider configurations. ADMIN role required.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - defaultProvider
                - parsingProvider
                - summariesProvider
              properties:
                defaultProvider:
                  type: string
                  enum: [ollama, gemini, claude]
                parsingProvider:
                  type: string
                  enum: [ollama, gemini, claude]
                summariesProvider:
                  type: string
                  enum: [ollama, gemini, claude]
      responses:
        '200':
          description: Configuration updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  updated:
                    type: object
                    properties:
                      defaultProvider:
                        type: string
                        enum: [ollama, gemini, claude]
                      parsingProvider:
                        type: string
                        enum: [ollama, gemini, claude]
                      summariesProvider:
                        type: string
                        enum: [ollama, gemini, claude]
        '400':
          description: Validation failed.
        '401':
          description: Unauthorized. Active session cookie required.
        '403':
          description: Forbidden. ADMIN role required.
...
  /api/profile/education:
    post:
      summary: Create education entry
      description: Creates a new education entry on the profile.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Education created.
        '401':
          description: Unauthorized.
  /api/profile/education/{id}:
    put:
      summary: Update education entry
      description: Updates the education and reconciles its bullets.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Education updated.
        '401':
          description: Unauthorized.
    delete:
      summary: Delete education entry
      description: Deletes the education entry.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Education deleted.
        '401':
          description: Unauthorized.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Session cookie generated by NextAuth.js